separate agent and plugin packages
