+++ Earthfile
@@ -4,18 +4,20 @@ FROM python:3.7-slim
 WORKDIR "/root/"
 
 all:
-    BUILD +deps
     BUILD +safety
-    BUILD +lint
-    BUILD +typecheck
-    BUILD +build
+    FOR PKG IN "prefix_list_agent" "prefix_list_agent_cli"
+        BUILD --build-arg PKG=$PKG +lint
+        BUILD --build-arg PKG=$PKG +typecheck
+        BUILD --build-arg PKG=$PKG +build-rpm
+    END
+    BUILD +build-swix
     BUILD --build-arg CEOS_VERSION="4.26.3M" +test
     BUILD --build-arg CEOS_VERSION="4.27.0F" +test
 
 deps:
     RUN apt update -qq && \
         apt upgrade -qqy && \
-        apt install -qqy git g++ swig libssl-dev
+        apt install -qqy git g++ swig libssl-dev zip
     RUN python -m pip install \
         --progress-bar off \
         --upgrade pip
@@ -30,9 +32,6 @@ deps:
     COPY Pipfile Pipfile.lock .
     RUN python -m pipenv sync --dev
 
-    COPY --dir share/ lib/ bin/ prefix_list_agent/ tests/ .
-    COPY LICENSE README.md pyproject.toml setup.cfg .
-
     LABEL org.opencontainers.image.source=https://github.com/wolcomm/eos-prefix-list-agent
     SAVE IMAGE --push ghcr.io/wolcomm/eos-prefix-list-agent/ci-deps:latest
 
@@ -41,48 +40,48 @@ safety:
     FROM +deps
     RUN --secret SAFETY_API_KEY python -m pipenv run safety check --full-report
 
-lint:
+src:
+    ARG --required PKG
     FROM +deps
+    COPY $PKG/ .
+    COPY LICENSE README.md .
+
+    LABEL org.opencontainers.image.source=https://github.com/wolcomm/eos-prefix-list-agent
+    SAVE IMAGE --push ghcr.io/wolcomm/eos-prefix-list-agent/ci-src:$PKG
+
+lint:
+    ARG --required PKG
+    FROM --build-arg PKG=$PKG +src
     RUN python -m pipenv run flake8 .
 
 typecheck:
-    FROM +deps
+    ARG --required PKG
+    FROM --build-arg PKG=$PKG +src
     RUN python -m pipenv run mypy --package prefix_list_agent
 
 sdist:
-    FROM +deps
+    ARG --required PKG
+    FROM --build-arg PKG=$PKG +src
     COPY --dir .git/ .
     RUN python -m pipenv run python -m build --sdist --outdir dist/
 
     SAVE ARTIFACT dist/*.tar.gz AS LOCAL dist/sdist/
 
-build:
+build-rpm:
     FROM fedora:31
 
     WORKDIR "/root"
 
     RUN dnf update -y
     RUN dnf install -y \
-            # python3 build deps
-            python3 \
-            python3-devel \
-            python3-pip \
+            # python build deps
+            python3 python3-devel python3-pip \
+            python2 python2-devel python2-pip python2-wheel \
             pyproject-rpm-macros \
             # rpm build tools
-            rpm-build \
-            rpm-devel \
-            rpmlint \
-            rpmdevtools \
+            rpm-build rpm-devel rpmlint rpmdevtools \
             # for switools
-            gcc-c++ \
-            swig \
-            openssl-devel
-
-    RUN rpmdev-setuptree
-
-    COPY packaging/rpm/eos-prefix-list-agent.spec ./rpmbuild/SPECS/
-    COPY packaging/swix/manifest.yaml .
-    COPY Pipfile Pipfile.lock .
+            gcc-c++ swig openssl-devel zip
 
     RUN python3 -m pip install \
             --disable-pip-version-check \
@@ -90,6 +89,7 @@ build:
             --no-warn-script-location \
             --user \
             pipenv
+    COPY Pipfile Pipfile.lock .
     RUN python3 -m pipenv lock --dev-only -r > requirements.txt
     RUN python3 -m pip install \
             --disable-pip-version-check \
@@ -98,19 +98,36 @@ build:
             --ignore-installed \
             -r "requirements.txt"
     
-    COPY +sdist/*.tar.gz rpmbuild/SOURCES/
+    ARG --required PKG
+    RUN rpmdev-setuptree
+    COPY packaging/rpm/$PKG.spec ./rpmbuild/SPECS/
+    COPY --build-arg PKG=$PKG +sdist/*.tar.gz rpmbuild/SOURCES/
 
     RUN SDIST="$(find rpmbuild/SOURCES/ -name '*.tar.gz' | head -n1)" && \
         VERSION="$(tar zxfO "$SDIST" --no-anchored PKG-INFO | grep '^Version:' | head -n1 | cut -d' ' -f2)" && \
-        rpmbuild -ba rpmbuild/SPECS/eos-prefix-list-agent.spec -D "_version $VERSION" && \
-        mkdir -p dist && \
-        SWIX="dist/eos-prefix-list-agent-${VERSION}.swix" && \
-        swix-create -i manifest.yaml "$SWIX" rpmbuild/RPMS/**/*.rpm
+        echo -n "$VERSION" > VERSION && \
+        rpmbuild -ba rpmbuild/SPECS/$PKG.spec -D "_version $VERSION"
 
-    SAVE ARTIFACT dist/*.swix AS LOCAL dist/swix/
+    SAVE ARTIFACT VERSION
+    SAVE ARTIFACT rpmbuild/RPMS/**/*.rpm AS LOCAL dist/rpm/
 
     LABEL org.opencontainers.image.source=https://github.com/wolcomm/eos-prefix-list-agent
-    SAVE IMAGE --push ghcr.io/wolcomm/eos-prefix-list-agent/ci-build:latest
+    SAVE IMAGE --push ghcr.io/wolcomm/eos-prefix-list-agent/ci-build-rpm:$PKG
+
+build-swix:
+    FROM +deps
+
+    RUN mkdir -p dist rpms
+
+    COPY --build-arg PKG="prefix_list_agent" +build-rpm/VERSION ./
+    COPY --build-arg PKG="prefix_list_agent" +build-rpm/*.rpm rpms/
+    COPY --build-arg PKG="prefix_list_agent_cli" +build-rpm/*.rpm rpms/
+
+    COPY packaging/swix/manifest.yaml .
+    RUN SWIX="dist/eos-prefix-list-agent-$(cat VERSION).swix" && \
+        python -m pipenv run swix-create -i manifest.yaml "$SWIX" rpms/*.rpm
+
+    SAVE ARTIFACT dist/*.swix AS LOCAL dist/swix/
 
 test-image:
     ARG --required CEOS_VERSION
@@ -129,7 +146,7 @@ test-image:
         pipenv
     RUN python3 -m pipenv sync --system
 
-    COPY pyproject.toml .
+    COPY prefix_list_agent/pyproject.toml .
     COPY tests/data/startup-config /mnt/flash/
     COPY --dir tests .
 
@@ -139,7 +156,7 @@ test-image:
     IF [ -n "$LOCAL_PACKAGE" ]
         COPY $LOCAL_PACKAGE /mnt/flash/
     ELSE
-        COPY +build/*.swix /mnt/flash/
+        COPY +build-swix/*.swix /mnt/flash/
     END
 
     RUN SWIX="$(basename $(find /mnt/flash -name '*.swix'))" && \
