+++ prefix_list_agent/setup.cfg
@@ -25,7 +25,7 @@ classifiers =
 zip_safe = False
 include_package_data = True
 packages = find:
-python_requires = >=3.7
+python_requires = ~=3.7
 scripts =
     bin/PrefixListAgent
