+++ prefix_list_agent_cli/setup.cfg
@@ -1,0 +1,44 @@
+[metadata]
+name = eos-prefix-list-agent-cli
+description = EOS CLI extensions for eos-prefix-list-agent
+long_description = file: README.md, LICENSE
+long_description_content_type = text/markdown
+author = Ben Maddison
+author_email = benm@workonline.africa
+url = https://github.com/wolcomm/eos-prefix-list-agent
+project_urls =
+    Documentation=https://wolcomm.github.io/eos-prefix-list-agent
+    Bug Tracker=https://github.com/wolcomm/eos-prefix-list-agent/issues
+keywords = bgp, irr
+license = MIT
+classifiers =
+    Development Status :: 4 - Beta
+    Environment :: Other Environment
+    Intended Audience :: Telecommunications Industry
+    License :: OSI Approved :: MIT License
+    Programming Language :: Python :: 2
+    Topic :: Internet
+    Topic :: System :: Networking
+
+[options]
+zip_safe = False
+include_package_data = True
+packages = find:
+python_requires = ~=2.7
+
+[flake8]
+max-line-length = 79
+doctests = True
+docstring-convention = pep257
+radon-max-cc = 10
+statistics = True
+count = True
+exclude = 
+    .git,
+    __pycache__,
+    build,
+    dist,
+    .venv,
+    .tox,
+per-file-ignores =
+    tests/*.py:S101,S404,S603
